+++ js/src/domains/planning/components/interventionMix/InterventionMixSummary.tsx
@@ -20,7 +20,6 @@ type Props = {
     setSelectedMix: (mix: number | null) => void;
     setMixName: (mixName: string) => void;
     setSelectedInterventions: (interventions: []) => void;
-    setMixApplied: (applied: boolean) => void;
 };
 
 const styles: SxStyles = {
@@ -41,7 +40,6 @@ export const InterventionMixSummary: FC<Props> = ({
     setSelectedMix,
     setMixName,
     setSelectedInterventions,
-    setMixApplied,
 }) => {
     const { formatMessage } = useSafeIntl();
     const { mutateAsync: createInterventionAssignment } =
@@ -76,7 +74,6 @@ export const InterventionMixSummary: FC<Props> = ({
         setMixName('');
         setSelectedInterventions([]);
         setSelectedMix(null);
-        setMixApplied(true);
     };
 
     const handleAssignmentCreation = async () => {
