+++ js/src/domains/planning/hooks/UseGetInterventionsPlan.tsx
@@ -7,19 +7,18 @@ import { InterventionPlan } from '../types/interventions';
 export const useGetInterventionsPlan = (
     scenarioId,
     selectedOrgUnits,
-    mixApplied,
 ): UseQueryResult<InterventionPlan[], Error> => {
     const params: Record<string, any> = {
         scenario_id: scenarioId,
         org_unit_ids: selectedOrgUnits,
     };
     const url = makeUrlWithParams(
-        '/api/snt_malaria/interventionassignments/grouped_by_mix/',
+        '/api/snt_malaria/interventionmixes/grouped_by_mix/',
         params,
     );
 
     return useSnackQuery({
-        queryKey: ['interventionPlans', scenarioId, mixApplied],
+        queryKey: ['interventionPlans', scenarioId, selectedOrgUnits],
         queryFn: () => getRequest(url),
         options: {
             staleTime: 1000 * 60 * 15,
@@ -27,7 +26,7 @@ export const useGetInterventionsPlan = (
             select: (data: InterventionPlan[]) => {
                 return data;
             },
-            enabled: Boolean(scenarioId) && mixApplied,
+            enabled: Boolean(scenarioId) && selectedOrgUnits.length > 0,
         },
     });
 };
